Rename content/fs to content/local
