+++ content/local/locks.go
@@ -1,4 +1,4 @@
-package fs
+package local
 
 import (
 	"sync"
