+++ content/local/store_linux.go
@@ -1,4 +1,4 @@
-package fs
+package local
 
 import (
 	"os"
