+++ content/local/store_test.go
@@ -1,4 +1,4 @@
-package fs
+package local
 
 import (
 	"bufio"
