+++ content/local/store_windows.go
@@ -1,4 +1,4 @@
-package fs
+package local
 
 import (
 	"os"
